requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.4
django-rest-framework==0.1.0
django_debug_toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
flake8==6.0.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
mixin==1.1
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
Black==23.7.0
